Split the 2232-line import-handler.ts import state machine into focused modules under services/import (line-parser, dependency-graph, conflict-resolver, an EntityImporter collaborator, and types), leaving import-handler.ts a 517-line orchestrator that still wraps the whole import in the same interactive transaction, with routes/import.ts and the public ImportHandler API unchanged.
