+++ docs/docs/project/changelog.md
@@ -67,6 +67,7 @@ The 0.5.0 cycle delivers the architecture-modularization roadmap (`notes/archite
 - Added `GET /api/projects/:projectId/assignable-users` (authorized by `manage_members`), returning the users who are not already members of a project. The "Add member" picker on the project detail page previously called the admin-only `/api/admin/users` endpoint, so it returned a 403 and showed an empty list to project owners and managers who were not system administrators; it now reads from the scoped endpoint and works for anyone who can manage the project's members.
 - Extracted the **claims** domain into a `ClaimRepository` (pure Prisma data access) and a `ClaimService` (orchestration, the relocated CASL authorization, and response mapping), reducing `server/src/routes/claims.ts` from 1559 lines with 33 direct `prisma.*` calls to 783 thin lines with zero. The claims routes already authorized through CASL, so every `can(...)` / `accessibleBy(...)` decision is preserved verbatim, along with the denormalized `claimsJson` rebuild, the claim-extraction and claim-synthesis queue jobs, the auto-created empty summary, and the subclaim cascade - a structural refactor with no behavior change.
 - Extracted the **world-state** domain into a `WorldStateRepository` (pure Prisma data access) and a `WorldStateService` (orchestration, the relocated CASL authorization, the demo / single-user read handling, and the object-reference gloss cleanup), reducing `server/src/routes/world.ts` from 1148 lines with 27 direct `prisma.*` calls to 372 thin lines with zero. The world routes already authorized through CASL, so every `can(...)` / `accessibleBy(...)` decision is preserved verbatim - including the existence-privacy distinction between forbidden and not-found, the anonymous / demo-mode read widening, the single-user default-user resolution, and the entity / event / time object-reference-to-text gloss cleanup on deletion - a structural refactor with no behavior change.
+- Split the 2232-line `server/src/services/import-handler.ts` import state machine into focused modules under `server/src/services/import/`: `line-parser.ts` (line parsing + validation), `dependency-graph.ts`, `conflict-resolver.ts` (conflict detection, resolution, id remapping, and cross-user resolution), `entity-importers.ts` (the per-type `EntityImporter` collaborator), and `types.ts` - leaving `import-handler.ts` a 517-line orchestrator. The pure parsing / graph / conflict logic became standalone functions; the per-type database writes moved to an `EntityImporter` that receives the transaction client on each call, so the entire import still runs inside the same single interactive `$transaction` (same `maxWait` / `timeout`, same write order). `routes/import.ts` and the public `ImportHandler` API are unchanged; behavior is identical.
 
 ### Fixed
